+++ sample-app/src/App.tsx
@@ -1,11 +1,14 @@
 import Container from "components/Container";
+import { I18nProvider, Localize } from "i18n";
 import React from "react";
 
 const App: React.FunctionComponent = () => {
     return (
-        <Container>
-            <div>hello world</div>
-        </Container>
+        <I18nProvider>
+            <Container>
+                <Localize id="header.title" />
+            </Container>
+        </I18nProvider>
     );
 };